requirements.txt
flask>=3.0.0
google-genai>=1.0.0
PyGithub>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
